Added music and door sound
I added a sound for when the player opens a door, and I added music to
the background. The music changes after floor 5. There are two of each
music file for a reason, some browsers cannot open MP3, some browsers
cannot open OGG, so the solution is to include both, and let the browser
sort out which it likes.