Lots of creature things
I have created a pool based creature spawning system. If the player is
between floor 1 and 5 (inclusive) it will select an enemy from the first
pool. 6 - 10 would have another pool and so on. This will allow the game
to still be random, but only select floor specific monsters.

I have also added a skeleton and a Bestiary (just for fun, it's in the
docs). There is a VERY basic hurt system (will be expanded later).